Implement more accessible memory structure for subjects (#1)
This is complete replacement for B3.Blackboard, but still
exposing interface to be possible to use it in B3 execution.

Using oneLine literal template function for better formatting
of invariant/warning messages.

Model value initialization now gives priority to function
specified with fallback to data passed in. That was it's
possible to apply transformation/validation on passed value.